+++ lib/components/common_touchable/TouchableEffect.dart
@@ -1,25 +1,54 @@
 enum TouchTypes {
+  ///
+  /// without any effect.
+  ///
   none,
+
+  ///
+  /// with atrubuite scaleValue, duration, lowerBound and upperBound.
+  ///
   scaleAndFade,
+
+  ///
+  /// with atrubuite scaleValue, duration, lowerBound and upperBound.
+  ///
   scaleAndUp,
+
+  ///
+  /// with atrubuite opacity value and duration.
+  ///
   opacity,
 }
 
 class TouchableEffect {
+  ///
   /// the typesof
+  ///
   TouchTypes type;
 
+  ///
   /// The value at which this animation is deemed to be dismissed.
+  ///
   final double scaleValue;
 
+  ///
   /// The value at which this animation is deemed to be dismissed.
+  ///
   final double opacity;
 
+  ///
   /// The value at which this animation is deemed to be dismissed.
+  ///
   final double lowerBound;
 
+  ///
   /// The value at which this animation is deemed to be dismissed.
+  ///
   final double upperBound;
+
+  ///
+  /// duration time for animation.
+  ///
   int duration;
 
   TouchableEffect({