+++ CHANGELOG.md
@@ -40,3 +40,16 @@
 ## 0.1.2
 * Added CommonDatePicker component.
 * Added more examples for Date picker.
+
+## 0.1.3
+* Added foreground decoration and handel custom foreground decoration.
+* Added new documentation with full examples and biggest update.
+
+## 0.1.6
+* Added more Comments and updated documentation.
+
+## 0.1.7
+* Added date picker time line and fixed border width problem.
+
+## 0.1.9
+* fixed date picker time line problem width problem.